+++ backend/src/main/java/com/fitsync/domain/user/entity/SocialProvider.java
@@ -1,4 +1,4 @@
-package com.fitsync.domain.user;
+package com.fitsync.domain.user.entity;
 
 public enum SocialProvider {
     GOOGLE, KAKAO, NAVER, GITHUB
