+++ backend/src/main/java/com/fitsync/domain/user/entity/UserGender.java
@@ -1,4 +1,4 @@
-package com.fitsync.domain.user;
+package com.fitsync.domain.user.entity;
 
 public enum UserGender {
     MALE, FEMALE, OTHER
