+++ backend/src/main/java/com/fitsync/domain/user/entity/UserType.java
@@ -1,4 +1,4 @@
-package com.fitsync.domain.user;
+package com.fitsync.domain.user.entity;
 
 public enum UserType {
     MEMBER, TRAINER, ADMIN
