+++ backend/src/main/java/com/fitsync/domain/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.fitsync.domain.user;
+package com.fitsync.domain.user.repository;
 
+import com.fitsync.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,4 +9,6 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
     public Optional<User> findByEmail(String email);
+
+    Boolean existsByEmail(String email);
 }
